Fixed several bugs and logic flaws

+++ SharpFly Packet Library/Packets/ClusterServer/Outgoing/QueryTickCount.cs
@@ -5,11 +5,11 @@ namespace SharpFly_Packet_Library.Packets.ClusterServer.Outgoing
 {
     public class QueryTickCount
     {
-        public QueryTickCount(int time, Socket socket)
+        public QueryTickCount(uint time, Socket socket)
         {
             OutgoingPacket packet = new OutgoingPacket(OpCodes.QUERY_TICK_COUNT);
             packet.Write(time);
-            packet.Write(DateTime.Now.Ticks);
+            packet.Write((uint)DateTime.Now.Ticks);
             packet.Send(socket);
         }
     }
